Fix style of Container Element
The container element previously, the images were not adequately centered in the container and the tect below the photographs were not in line. I have changed this by editing the html tag for the text element below the photographs to adequately style it. I have also included a minimum width for the container and changed the width of the images in order to properly center everything.
I have also changed the testimonial div by including another div for the text in order to style the text in css.