Fix the position of the footer div.
The footer was previously not sticking to the bottom of the page. I have edited the css for the footer div to remove the syntax postion: fixed and adjusted the height, and the padding at the top and bottom of the footer. The footer is now visible and stationery at the bottom of the page.